% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcd_data.R
\name{lcd_data}
\alias{lcd_data}
\title{A Data Download Function}
\usage{
lcd_data(x, fun)
}
\arguments{
\item{x}{x must equal a vector of LCD stations.}

\item{fun}{The function parameter must equal "download." If else, the function will not run. This parameter indicates that you wish to download data.}
}
\description{
This function downloads meteorological station data from NOAA's Local Climatological Data (LCD). Additionally this function utilizes the \code{lcd} function from the \code{rnoaa} package. The purpose of this function is to create a tibble of station data over a 10 year period.
}
\details{
The function is designed to download data for the years 2005 to 2014, however the code can be adapted to the user's data needs. The code iterates through a list of LCD stations to create a list of tibbles for each station for the 10 year period. For more information on the \code{lcd} function, please refer to the \code{rnoaa} package.
}
\seealso{
\code{\link[rnoaa]{lcd}} for more information on the function utilized in \code{lcd_data}.
}
